# PERF: cythonize kendall correlation