- [x] closes #28329 
- [x] tests added / passed
- [x] Ensure all linting tests pass, see [here](https://pandas.pydata.org/pandas-docs/dev/development/contributing.html#code-standards) for how to run them
- [x] whatsnew entry

ASV's. 

       before           after         ratio
     [a6bc6ecc]       [1eb86441]
     <perf-cythonize-kendall~2^2>       <perf-cythonize-kendall>
-             81M            57.6M     0.71  stat_ops.Correlation.peakmem_corr_wide('kendall')
-      1.41±0.02s          362±7ms     0.26  stat_ops.Correlation.time_corr_wide_nans('kendall')
-      1.57±0.04s          398±4ms     0.25  stat_ops.Correlation.time_corr_wide('kendall')
-        34.6±4ms      8.75±0.06ms     0.25  stat_ops.Correlation.time_corr('kendall')

SOME BENCHMARKS HAVE CHANGED SIGNIFICANTLY.
PERFORMANCE INCREASED.