@jbrockmendel 2021-01-20
couple nitpicks, generally looks good.  i take it we have sufficient testing?